Change Project structure for api end points
