+++ src/recipe_management_project/following_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class FollowingAppConfig(AppConfig):
+    name = 'following_app'
