+++ src/recipe_management_project/recipe_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class RecipeAppConfig(AppConfig):
+    name = 'recipe_app'
